pew pew dat mesh

+++ Dependencies/Include/BoundingVolumes.h
@@ -120,7 +120,7 @@ public:
     bool checkCollision(Triangle* t) override;
     bool checkCollision(Capsule* col) override;
 
-    bool isInside(glm::vec3 point);
+    bool isInside(glm::dvec3 point);
 
     Triangle(glm::vec3 v0, glm::vec3 v1, glm::vec3 v2, glm::vec3 norm);
     void toString() override;
